Move agent gen out of master stack+vpcazs
* Agent generator code moved out of the master stack
* VPC availability zone function moved into vpc stack,
  which is where it should have been to begin with
